+++ src/nns/macros/mod.rs
@@ -1,0 +1,12 @@
+#[macro_use]
+mod cache_errors;
+#[macro_use]
+mod leaf_requests;
+#[macro_use]
+mod network;
+#[macro_use]
+mod reports;
+
+#[cfg(test)]
+#[macro_use]
+mod test_helpers;